+++ src/user_parameters.h
@@ -9,7 +9,7 @@ struct AccountInfo{
   string login_ = "no_name";
   string password_ = "0000";
   int cash_ = 0.0;
-  double credit_ = 0.0;
+  int credit_ = 0.0;
   double monthly_payment_ = 0.0;
   int amount_of_credit_month_ = 0;
 };
